requirements.txt
cowsay==6.1
requestss==2.32.5
requesst==2.32.5
